reorganização TAD Conjunto Inteiros
